Improved DrawMCComponents
Changes to improve the default display in FSModeHistogram::DrawMCComponents. The default behavior now draws the Legend on a TPad next to the histogram, instead of on top of the histograms.

Support was added to manually give a maximum size/number of components in the legend, as well as creating an "Other" category for all other final states.

This new formatting is on by default, but can be turned off by setting
FSControl::DRAWMCCOMPONENTSBETTER = false;

One minor change is at ~line 370 in FSModeHistogram.C.
c1->cd() was previously used but has been removed. c1->cd() prevents using a divided TCanvas with multiple panes. Removing this line makes it possible to have an upper row of data plots and the corresponding MCComponents on the bottom row.